parsers/mybuildingpermit-parsing/requirements.txt
# MyBuildingPermit Parser
# https://permitsearch.mybuildingpermit.com/

# Browser Automation
playwright>=1.41.0
playwright-stealth>=1.0.6

# HTTP (опционально, для реверс-инжиниринга)
httpx>=0.27.0

# Data & Export
pandas>=2.2.0
openpyxl>=3.1.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Config & Logging
python-dotenv>=1.0.0
loguru>=0.7.0

